include support for "author intended blank page"
According to the dissertaion specs, blank pages need to
include a message. The original template didn't (ever? always?)
print a message on blank pages.  There is now always a message
printed.  You can change the message by searching for "blank page"
in the style file and adjusting as preferred.  This inclusions
introduced a dependency on \phantomsection.  If you're not
already including the hyperref package, then you'll want to look
at the sample thesis.tex to see the work-around.